Refactor email verification jobs
